// src/app/api/model/error-response.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Negative response for an API call
 */
export interface ErrorResponse { 
    /**
     * The specific error code defined
     */
    error_code?: string;
    /**
     * The description of the error
     */
    error_description?: string;
    /**
     * The request http status
     */
    status?: number;
    /**
     * The date and time when the error occurred
     */
    timestamp?: string;
}